Small improvements

+++ flash.sh
@@ -2,12 +2,14 @@
 
 BUILD_ONLY=false
 VERBOSE=false
-NO_HTML=true
+NO_HTML=false
+TIZEN_IP="192.168.41.64"
+GBS_ROOT="$HOME/GBS-ROOT"
+BUILD_NAVIT=true
 
 # getopt section
-TEMP=`getopt -o b,v --long build-only,verbose -n 'flash.sh' -- "$@"`
+TEMP=`getopt -o b,v,h,i:,g:,n --long build-only,verbose,help,tizen-ip:,gbs-root:,no-navit -n 'flash.sh' -- "$@"`
 eval set -- "$TEMP"
-
 while true; do
     case "$1" in 
     --build-only)
@@ -18,6 +20,23 @@ while true; do
         shift;
         VERBOSE=true
         ;;
+    -h|--help)
+        shift;
+        echo "usage $0 -h -b -v"
+        exit 0
+        ;;
+    -i|--tizen-ip)
+        TIZEN_IP=$2
+        shift 2;
+        ;;
+    -g|--gbs_root)
+        GBS_ROOT=$2
+        shift 2;
+        ;;
+    -n|--no-navit)
+        BUILD_NAVIT=false;
+        shift ;
+        ;;
     --)
         shift;
         break;
@@ -25,10 +44,8 @@ while true; do
     esac
 done
 
-TIZEN_IP="192.168.41.64"
 NXE_VERSION=0.1.1
 NAVIT_VERSION=
-GBS_ROOT="$HOME/GBS-ROOT"
 
 red=`tput setaf 1`
 green=`tput setaf 2`
@@ -47,20 +64,29 @@ if [ "$BUILD_ONLY" = false ]; then
 
     # remove nxe and navit
     ssh root@$TIZEN_IP zypper -n -q rm nxe > /dev/null 2>&1
-    ssh root@$TIZEN_IP zypper -n -q rm navit > /dev/null 2>&1
+    ssh root@$TIZEN_IP zypper -n -q rm nxe-debugsource > /dev/null 2>&1
+
+    if [ "$BUILD_NAVIT" = true ]; then
+        ssh root@$TIZEN_IP zypper -n -q rm navit > /dev/null 2>&1
+        ssh root@$TIZEN_IP zypper -n -q rm navit-debugsource > /dev/null 2>&1
+        ssh root@$TIZEN_IP zypper -n -q rm navit-debuginfo > /dev/null 2>&1
+    fi
 fi
 
-# build navit
-echo "${red}Building navit${reset}"
-if [ "$VERBOSE" = true ]; then
-    try gbs build -A i586 --spec navit_qt5.spec --include-all 
-else
-    try gbs build -A i586 --spec navit_qt5.spec --include-all > /dev/null 2>&1
-fi
-if [ "$BUILD_ONLY" = false ]; then
-    try scp $GBS_ROOT/local/BUILD-ROOTS/scratch.i586.0/home/abuild/rpmbuild/RPMS/i686/navit-0.5.0.6011svn-1.i686.rpm root@192.168.41.64:/root
-    try scp $GBS_ROOT/local/BUILD-ROOTS/scratch.i586.0/home/abuild/rpmbuild/RPMS/i686/navit-debuginfo-0.5.0.6011svn-1.i686.rpm root@192.168.41.64:/root
-    try scp $GBS_ROOT/local/BUILD-ROOTS/scratch.i586.0/home/abuild/rpmbuild/RPMS/i686/navit-debugsource-0.5.0.6011svn-1.i686.rpm root@192.168.41.64:/root
+if [ "$BUILD_NAVIT" = true ]; then
+    # build navit
+
+    echo "${red}Building navit${reset}"
+    if [ "$VERBOSE" = true ]; then
+        try gbs build -A i586 --spec navit_qt5.spec --include-all 
+    else
+        try gbs build -A i586 --spec navit_qt5.spec --include-all > /dev/null 2>&1
+    fi
+    if [ "$BUILD_ONLY" = false ]; then
+        try scp $GBS_ROOT/local/BUILD-ROOTS/scratch.i586.0/home/abuild/rpmbuild/RPMS/i686/navit-0.5.0.6011svn-1.i686.rpm root@192.168.41.64:/root
+        try scp $GBS_ROOT/local/BUILD-ROOTS/scratch.i586.0/home/abuild/rpmbuild/RPMS/i686/navit-debuginfo-0.5.0.6011svn-1.i686.rpm root@192.168.41.64:/root
+        try scp $GBS_ROOT/local/BUILD-ROOTS/scratch.i586.0/home/abuild/rpmbuild/RPMS/i686/navit-debugsource-0.5.0.6011svn-1.i686.rpm root@192.168.41.64:/root
+    fi
 fi
 
 # build nxe
@@ -78,7 +104,9 @@ fi
 
 if [ "$BUILD_ONLY" = false ]; then
     # install both
-    try ssh root@$TIZEN_IP rpm -ivh /root/navit*
+    if [ "$BUILD_NAVIT" = true ]; then
+        try ssh root@$TIZEN_IP rpm -ivh /root/navit*
+    fi
     try ssh root@$TIZEN_IP rpm -ivh /root/nxe*
 fi
 
@@ -86,6 +114,9 @@ if [ "$NO_HTML" = false ]; then
     # example
     cd Navigation/NXE/exampleapp
     try ./make_xpk.sh app ~/mykey.pem
-    try scp app.xpk root@$TIZEN_IP:/home/app
-    try ssh root@$TIZEN_IP chown app /home/app/app.xpk
+    try scp app.xpk app@$TIZEN_IP:/home/app
+
+    # stop xwalk
+    try ssh app@$TIZEN_IP pkgcmd -u -n cjhbbeknomcehebnhobpolialjjnalad -q
+    try ssh app@$TIZEN_IP pkgcmd -i -t xpk -p /home/app/app.xpk -q
 fi
